implement for pistols

+++ planetary_interface/src/interfaces/one_on_one.cairo
@@ -2,8 +2,8 @@ use starknet::ContractAddress;
 
 #[starknet::interface]
 pub trait IOneOnOne<TState> {
-    fn create_match(ref self: TState, p1: ContractAddress, p2: ContractAddress, playlist_id: u32) -> u32;
-    fn settle_match(ref self: TState, match_id: u32) -> Status;
+    fn create_match(ref self: TState, p1: ContractAddress, p2: ContractAddress, playlist_id: u128) -> u128;
+    fn settle_match(ref self: TState, match_id: u128) -> Status;
 
 }
